Deleted the mermaid.js configuration file
The mermaid.js file has been deleted from the Frank.Blazor.Mermaid project. This configuration file was primarily managing the configuration of logs and charts. This change could potentially affect the functionality of the application. The reason for this deletion and its possible replacement or solution should be addressed in the subsequent commits.